Drop IFluidHTMLView from DO, start conversion to react
